Move VariableFilter to its own file
